chore: ButtplugDevice now an abstract class, added DeviceAddedMessage
Made ButtplugDevice an abstract class, just to move the Name property
back a level and add logger setup. Implemented the DeviceAddedMessage
message and removed the DeviceAddedEvent from being sent to outside
handlers, so that all lifetime of device objects can be managed by the
core buttplug service. All interaction now done via message passing,
as it was originally meant to be.