test: Move try block back to type fetching portion
Problem is in calling Assembly.GetTypes(). The exception is thrown
because appveyor doesn't seem to have the assembly we're looking for
bluetooth symbols in, and so it'll drop all those symbols. We should
be able to null-test for them in the resulting list that we get with
the exception though. See
http://haacked.com/archive/2012/07/23/get-all-types-in-an-assembly.aspx/
for more info.