chore: Refactor CreateDeviceAsync to be generic
Instead of having a factory plus a info struct per bluetooth device
type, we now just device a set of filters and a factory function. This
way we can use the same creation function for all devices, and split
API versions between managers where needed (for instance, >= 15063 and
< 15063) while still using the same device code. Also contains some
resharper refactoring work as resharper is fairly insistent on what it
wants to do.

Fixes #8