fix: Fix crash on connect when device powers up after scan started
Turned the fleshlight launch on after hitting "Start scanning" in the
GUI, and the program threw System.Exception. Turns out that windows
can try to connect to devices before its completely ready. If this
happens, just log and stop that round of connection, go back to
waiting for the next advertisement. It usually resolves itself after
2-3 more advertisements.