chore: Remove Check funcs and just throw fucking exceptions already
When in Rome, do not cover your ears and scream "I AM NOT IN ROME".

Was using some odd combination of Check() and prayer to try and deal
with invalid values in messages, in order to simulate rust's
"constructors are just a function named new that can return a Result"
idea. Moved logic into property setters and just throw
ArgumentException if something goes wrong. This allows us to break out
of the json parser if a message has incorrect values, and will also
support outside developers in a way they expect once we get to that
point. May add static "CreateMessage" members for messages that just
return Option<> if I really need that functionality for myself down
the road.