feat: Make logmanagers per-service-instance
Create the log manager in the ButtplugService, and pass it down
through the system. Probably not the best way to do this, but it works
for now and I'd rather not create monolithic singleton managers before
I have to. That said, might be cleaner to just have ButtplugService
create some sort of token ID and pass that around in the future.

Fixes #41