feat: Move from uhttpsharp to HttpListener
uhttpsharp was causing thread deadlocks due to using a version of
LibLog that was older than what Buttplug uses. Also, we were having
problems getting uhttpsharp to stop listening on sockets. Since the
Kiiroo emulator is so simple, we can just use asynchronous
HttpListener for the 2 REST verbs we need. This will need more tests
now that we're handling it ourselves though.

Fixes #42
Fixes #39
Affects #3