feat: Abstract Bluetooth APIs to make device protocols generic
Bluetooth device access was tied directly to the UWP APIs. This patch
abstracts the calls and factories into interfaces, so they can be
plugged in as needed. This will allow us to make new backends for
Xamarin, as well as test Bluetooth protocols in environments where we
don't have access to hardware/APIs.

Fixes #35