chore: Remove assembly reflection, make default bluetooth info list
Assembly reflection has been a constant problem of weird reflection
load exceptions depending on what assemblies are on a system. It's not
like we'll be adding devices to the default list that often anyways.
Tests added to make sure all default devices show up in the new
BluetoothSubtypeManager default class.

Fixes #56
Fixes #55