feat: Remove Languageext and start using resharper annotations
LanguageExt was really nice for either/option, but the library
dependencies were tough to deal with (ValueTuple fighting with
PresentationFramework/LanguageExt, etc), and it's also just a big
library that changes the langauge a lot but in non-ergonomic ways.
Now using Resharper static analysis plus more LINQ calls to see if I
can work without LanguageExt in the code.

Fixes #49