fix: Make sure characteristics instantiate in GUID order
The LINQ query used to come up with the characteristic array wasn't
order characteristics in any way, so we got back a non-deterministic
array of characteristics that, until now, mostly worked. Fixed that to
order by GUID for the moment, though it should be ordered by the GUID
characteristic array in the devices info object.