feat: Add server name settings and max ping time to server init
Synchronizing handshake phase messages between spec and reference
implementation of server. We now set a server name and a maximum ping
time on startup. We also expect that RequestServerInfo will always be
the first message to come across the line, and error if it is not.

Updated tests to reflect this, though we handle the RequestServerInfo
send in the constructor of TestServer so that we don't have to do it
on every initialization there.

Fixes #80
Fixes #82