fix: Fix dealing with device counter increment in threads
.Net uses an M:N threading model to back its async system, meaning the
threads pool is opaque to us, but it's still there. We need to add
checks like we're threading this ourselves. In this case, if 2 devices
connected at almost exactly the same time, we could have a collision
where they both tried to use the same device index. This would cause a
crash when the keys collided in the Dictionary.

Fixes #106