feat: Add Lovense Edge support, break Lovense Info into revisions
The Lovense Edge uses the same commands as the other lovense
revisions, but requires new service/characteristics UUID and a name.
To support all 3 revisions of Lovense hardware, 2 IBluetoothDeviceInfo
classes are added. This should hopefully let us support everything we
have device names for at the moment. Only released toy we don't
support is the Ambi, which we need a device name for.

Fixes #105