Adding HTML landing page on /index.html
The web socket library can be used as a HttpServer too.
This will give us a test page we can point browsers to to test SSL issues etc.

In the future, it'd be cool to present syncydink here, so that users can test with just a browser (no additional media required).