Adding ping timeouts
See issues #81 (Ping keeps the connection alive, terminating unresponsive clients).

The websocket default timeout is 100ms, but this can overriden in the %APPDATA%\Buttplug\config.json file:

```json
{
  "buttplug": {
    "server": {
      "port": "12345",
      "secure": "True",
      "maxPing": "0"
    }
  }
}
```

Note that 0 disables the ping timer.

On the way, we discovered that all of the WebSocket were sharing the same ButtplugService (meaning that the first connection to miss its ping blocked all other and future connections). Some refactoring has been done to mitagate this.

The Kiiroo emulator doesn't have a concept of a constant connection between the video player and the server, so the timeout has been hardcoded to 0.