fix: async all the asyncable things
Using .GetAwaiter().GetResult() can deadlock in some instances,
depending on the threadpool resources we've got, etc... This means we
can see hangs on places like CI, where we may not have the resources
to spin up a large threadpool. Async/await everything so that the
threadpool can take care of things in whatever way it wants.