fix: Fix websocketclient hangs on appveyor
WebsocketClient was hanging on appveyor due to a race that would cause
CloseAsync to wait on a response from a server that didn't exist
anymore. This patch changes WebsocketClient to use CloseOutputAsync,
which is a "fire and forget" version that does not wait on a response.
It also adds a catch for the cancellation token activation that may
stop the ReceiveAsync call in the reader thread after this.