Adding an 'ignore cert errors' flag to the WSClient
This sets the ignore subject and ignore unknown certificates flags on WebSocket4Net.

Unfortunatly our WebSocketServer library does not allow for a X509Chain to be provided and WebSocket4Net requires
the chain to be valid (even if the root is self signed). To avoid this breaking us further, the cert generation now skips the CA and only gnerates a single self-signed cert. For upgrade purposes, any existing CA+Cert combos wil be removed.

Fixes #316