Correcting the handing of old messages arriving at a new parser
I'm not sure if this is actually possible in the wild, since the client/server relationship is resticted to server >= client.
In testing, I serialised DeviceAdded to Version0, then deserialised it, causing an error since it didn't match the current object version. At this time, the client does not send any messages that the server would choke on, but this builds that in for future cases.

Also more tests and some ReShaper work.