ButtplugWSClient now requests the device list on connect
This means that during Connect() any devices already connected to the buttplug server will cause DeviceAdded events to be fired. This will include any devices that may have already been seen in the case of a reconnect (if the connection has dropped anything could have happened to the server, therefore the device list is purged and all devices are readded from scratch). As this is now called for you, the RequestDeviceList() method has been removed.

The internal device list is now accessible via a C# style getter Devices rather than the previous getDevices() method.

The nextMsgId getter has also been renamed to NextMsgId for consistency.