fix: Catch Cryptographic Error and throw appropriate error message
Code was throwing an uncaught expection on cert gen when the cert
can't be written to disk, which would bubble beyond exception handlers
and crash the application. Catch the exception and tell the user to
turn off SSL. Also provide forum info on about tab so users can
contact devs for support.

Fixes #372