feat: Adding more support BLE subscribtions
This change adds support for subscribing to specif characteristics, as
well as improving the event handling so that the source characteristic
and time of receipt are also passed to the subscriber.

This is required for the no-op subscription performed by the gen1 Kiiroo
devices, as well as for future device input support.