feat: Support for the Gen1 Kiiroo Onyx
The Onyx takes a target 'position' of 0-4 that controls a number of
gripping rings. In order to support time based movement, we have to
calculate which position we should be at on a regular basis and update
the position on the device accordingly.

The device also needs to be sent the position once every so often or it
will relax back to it's base position when idle.

An emulator for Fleshlight*Cmd support is also included, mapping speed
and position delta into a time value.