chore: Move Client to WebsocketListener, Start making the client APIs idiomatically async
WebsocketListener now comes with a client, meaning we can chuck the
mess that is Websocket4Net. Also start post-fixing "async" onto
functions that are async, and allowing them to take cancellation
tokens where possible.

This should also get rid of the "Not GET request" error on the
websocket, as that is no longer an error in WebsocketListener. We may
get some new error, but that should get a new bug to go with it.

Fixes #279, #283