feat: Receiving errors from a server should throw exceptions in client
Instead of just passing back an error message and expecting a client
to check for it, we should throw exceptions whenever we receive error
messages from the server. This should happen in both embedded and
remote connectors for the client, meaning we may need to SetExceptions
in connectors instead of just doing SetResult.