feat: Use device types and device message sending functions to ButtplugClientDevice
ButtplugClientDevice assumed that the user would build message objects
to send. This is bad. The user shouldn't have to do that. We have an
API for a reason, and it's not for the user to build raw messages.

Also, the user should never see message types as strings. They should
always be Types, because then we're at least constrained to the
universe of types in the program, versus "any string ever".

Fixes #521
Affects #524 (but doesn't fix it because json conversion gets complicated)