feat: Autoscan for DeviceSubtypeManagers on first call to StartScanning
Instead of requiring users to add each Subtype Manager, just scan for
all of them during the first StartScanning call. If none are found,
throw.

Removes the test for no SubtypeManagers, as we'll be including the
SerialPort subtype manager with the core library, so there shouldn't
be a point where we have no managers.

Fixes #526