feat: Adding support for more MagicMotion devices
This completes the set by supporting vibrate on the MagicMotion Magic
Wand and MagicMotion Magic Kegel Twins.

The Kegel Twins and Master share names, so more work is required to
individually identify the devices. This isn't important right now since
the hardware difference is the presure sensor, which we have no support
for yet.

The Magic Wand has a thermometer and heating element, but I haven't
decoded the controls for those yet.