chore: Add motion-first-sent status to all protocols
Protocols all start with "last device state variables" that use
primitive, non-nullable types. This means if the first command we send
to a vibration device is 0, then the device thinks it's already at
zero and doesn't sent it. Added boolean values that protocols can
change in order to start skipping repeated commands after the first
one is sent.

Fixes #564