fix: Fixing a couple of issues that blocked device reconnection
In the Bluetooth Manager, devices that connected were added to the seen list, so on reconnection were ignored unless the scanning was toggled on and off again.

In the ButtplugDevice code, we were defereeing to the inner device object for a connection status, but I was seeing the same device object being reused after power cyclying the device, which caused the device manager to error with duplicate device before telling the client that the device was back.