fix: Only add BLE devices to seen list when they lack a matching config
Devices were getting added to the seen list in places they shouldn't
be, included after having successfully connected. In reality, devices
should only be added to the seen list when they are seen but do not
match any of the devices we expect to see in the
DeviceConfigurationManager.

Fix originally introduced in #599, cleaned up here.