feat: Adding more LiBo devices
This change adds support for the LiBo LaLa, LuLu, Karen and Lottie, as well as
mapping the product name of the ELLE (aka Whale).

The Karen is mapped and registers with only the StopDeviceCommand since it
doesn't have a vibe.

The LaLa has a suction tip which is mapped as the second vibrator: this controls
the 3 suction patterns, which would appear to only differ in on/off frequency,
somewhat emulating intensity for a actuator that has no other clear means of
providing intensity.

The ELLE is actually mapped to Elle since the all-caps looks like a mistake.
The vibe, which has 3 levels of intensity, is mapped to index 0 and the estim
output is mapped to index 1. This may be controversial, but the device is not
prickly at all, and does give vibe like feels.