build: Add a disabled version of the IntifaceCLI project
Just links files, and expects the buttplug-csharp and
intiface-cli-csharp repos to be in the same directory. Seems to
compile though, and means we can work on the CLI in relation to the
library without having to bind the two at the version level. May need
some sort of git hook to make sure IntifaceCLI project is never
checked in loaded, though CI should fail if that happens anyways.