Added Nova 2 and protocol for Nova 2 and Chorus fix
Added support for Nova 2 and a fix to ensure Chorus and Nova 2 reach full speed.  There may be other devices that could benefit, I only have the two to test with.  I am not 100% certain this is the optimal data to send to control the Nova 2 and Chorus, however I am certain it is working close to optimally.
(Also Chorus apparently now works in Windows 10 if the device is paired to via Bluetooth.)